Rebranding
We are not affiliated, associated, authorized, endorsed by, or in any way officially connected with WhatsApp, or any of its subsidiaries or its affiliates. The official WhatsApp website can be found at https://www.whatsapp.com.

The name WhatsApp as well as related names, marks, emblems and images are registered trademarks of their respective owners. (Not us!)